requirements.txt
google-play-scraper
google-generativeai
dotenv
pandas
transformers
torch